requirements.txt
# Core dependencies
aiohttp==3.9.1
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Web Scraping (Discord via browser)
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Discord API (deprecated - use web scraping instead)
# discord.py==2.3.2  # DO NOT USE - violates compliance

# HTTP/Utilities
requests==2.31.0
colorama==0.4.6

# Solana
solana==0.30.2
solders==0.18.1
base58==2.1.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
